Add content management functionality to ZauberCMS
This commit adds numerous features related to content management in the ZauberCMS application. It includes the addition of ContentMapper, GetContentsCommand, and SaveContentCommand classes along with their respective handlers, GetContentsHandler and SaveContentHandler. Also, several updates are made to the web interface, especially to the Admin section for the management of content and their types.