Add new contact form and views, enhance content type system
New ContactForm and ContactView components were added as part of enhancing the StarterSite. AnotherView and WebsiteView were renamed to TextPageView and HomeView respectively and were updated with new content. This update also includes changes to the content type system; new boolean properties, RootOnly and IsElementType, were added to GetContentTypesCommand and ContentType model respectively, allowing to filter content types that can be placed at the root or be categorised as an Element type. All in all, these changes provide clear and structured content handling.