Add user edit tracking and published status to content
This commit introduces two key changes to the content infrastructure. Firstly, it adds a 'LastUpdatedById' field to the 'ZauberMedia', 'ZauberContentTypes', and 'ZauberContent' tables, allowing for tracking who last edited a piece of content. Secondly, a 'Published' field is added to the 'ZauberContent' table to enable recording of content publication status. Both changes involve necessary migration updates for SqLite and SQL Server databases.